requirements.txt
amqp==2.6.0
asgiref==3.2.7
beautifulsoup4==4.7.1
billiard==3.6.3.0
bleach==3.1.0
celery==4.4.5
certifi==2019.3.9
chardet==3.0.4
Django==3.0.6
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-contrib-comments==1.9.1
django-crispy-forms==1.9.1
django-timezone-field==4.0
dnspython==1.16.0
eventlet==0.25.2
filebrowser-safe==0.5.0
future==0.18.2
grappelli-safe==0.5.2
greenlet==0.4.16
idna==2.8
importlib-metadata==1.6.1
kombu==4.6.10
monotonic==1.5
numpy==1.17.0
oauthlib==3.0.1
pandas==1.1.0
Pillow==6.0.0
psycopg2==2.7.5
psycopg2-binary==2.8.2
python-crontab==2.5.1
python-dateutil==2.8.0
pytz==2018.4
redis==3.2.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
tzlocal==1.5.1
urllib3==1.25.3
vine==1.3.0
virtualenv==16.0.0
webencodings==0.5.1
zipp==3.1.0
